Refactored Package Details
Packages: model, utility, and ui have been added using the refactoring methods that were taught in the lab on Nov 1st.
