+++ app/src/main/java/ca/ualberta/cs/cmput301f18t19/hada/hada/model/CareProvider.java
@@ -8,7 +8,7 @@
  * Copyright (C) 2018  Austin Goebel, Anders Johnson, Alex Li,
  * Cristopher Penner, Joseph Potentier-Neal, Jason Robock
  */
-package ca.ualberta.cs.cmput301f18t19.hada.hada;
+package ca.ualberta.cs.cmput301f18t19.hada.hada.model;
 
 
 /**
