+++ app/src/main/java/ca/ualberta/cs/cmput301f18t19/hada/hada/model/Problem.java
@@ -9,7 +9,7 @@
  * Cristopher Penner, Joseph Potentier-Neal, Jason Robock
  */
 
-package ca.ualberta.cs.cmput301f18t19.hada.hada;
+package ca.ualberta.cs.cmput301f18t19.hada.hada.model;
 
 import java.time.LocalDateTime;
