+++ app/src/main/java/ca/ualberta/cs/cmput301f18t19/hada/hada/model/RecordListController.java
@@ -1,6 +1,8 @@
-package ca.ualberta.cs.cmput301f18t19.hada.hada;
+package ca.ualberta.cs.cmput301f18t19.hada.hada.model;
 
 import java.util.ArrayList;
+
+import ca.ualberta.cs.cmput301f18t19.hada.hada.utility.Listener;
 
 public class RecordListController {
