+++ app/src/main/java/ca/ualberta/cs/cmput301f18t19/hada/hada/ui/MainActivity.java
@@ -8,10 +8,12 @@
  * Copyright (C) 2018  Austin Goebel, Anders Johnson, Alex Li,
  * Cristopher Penner, Joseph Potentier-Neal, Jason Robock
  */
-package ca.ualberta.cs.cmput301f18t19.hada.hada;
+package ca.ualberta.cs.cmput301f18t19.hada.hada.ui;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
+import ca.ualberta.cs.cmput301f18t19.hada.hada.R;
 
 public class MainActivity extends AppCompatActivity {
