+++ app/src/test/java/ca/ualberta/cs/cmput301f18t19/hada/hada/model/PatientListControllerTest.java
@@ -1,6 +1,9 @@
-package ca.ualberta.cs.cmput301f18t19.hada.hada;
+package ca.ualberta.cs.cmput301f18t19.hada.hada.model;
 
 import org.junit.Test;
+
+import ca.ualberta.cs.cmput301f18t19.hada.hada.model.Patient;
+import ca.ualberta.cs.cmput301f18t19.hada.hada.model.PatientListController;
 
 import static org.junit.Assert.*;
