+++ app/src/test/java/ca/ualberta/cs/cmput301f18t19/hada/hada/model/RecordListControllerTest.java
@@ -1,6 +1,9 @@
-package ca.ualberta.cs.cmput301f18t19.hada.hada;
+package ca.ualberta.cs.cmput301f18t19.hada.hada.model;
 
 import org.junit.Test;
+
+import ca.ualberta.cs.cmput301f18t19.hada.hada.model.Record;
+import ca.ualberta.cs.cmput301f18t19.hada.hada.model.RecordListController;
 
 import static org.junit.Assert.*;
